Renames node creation function and moves it to utils.
